Change of vamp/cult telepathy styles (#767)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Just a slight change in styles of vampire and cultists telepathy text
from godawful yellow to something like this


![scr1](https://github.com/user-attachments/assets/ae3ac735-7261-4fc7-86bd-eb82a0b573a3)

![scr2](https://github.com/user-attachments/assets/10898fe9-3697-4501-bda4-6cf78f3a268e)

<!-- Describe 
The Pull Request. Please be sure every change is documented or this can
delay review or prevent the PR from being merged! -->

## Why It's Good For The Game
Old yellow text was super easy to miss in the countless action messages
that had almost the same color as telepathy. Now it would be easier to
read it and determine from who message comes from, since it have the
color of person's voice.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->