(FEATURE) Add Skill Descriptions (#686)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

What it says! Adds a ? button next to your skill level. When pressed, it
shows a small generic description of what the skill does in game and how
it can be trained.


![image](https://github.com/user-attachments/assets/f52836b1-99a2-4623-8573-7df1bd5f6e35)

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
This is a quality of life for new players so things aren't so esoteric.
It helps to know that combat skills affect parry and how likely you are
to hit.
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->