Smithy Remap (By CocoKailey) (#759)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Reworks the weapon and armor smith building.
- Gives them a backpack and cart for improved hauling.
- Gives them more space.
- Gives them more efficient tool layouts.
- Weapon smith has a north and west table
- Armor smith has an east and south table
- Weapon smith's room is accessed by a staircase on the exterior, it is
no longer connected directly to the store.


![01](https://github.com/user-attachments/assets/3ea035ec-6a95-4a91-b79d-effff5ae1c64)

![02](https://github.com/user-attachments/assets/c9ba5389-7f27-4bca-a4c7-2178ae213a09)

![03](https://github.com/user-attachments/assets/8a2963f2-81a4-4411-b2d0-960ddb2afde9)

![04](https://github.com/user-attachments/assets/bd46f9bb-a0cc-4a13-91e6-a24824975990)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
It helps the blacksmiths bee a little bit more efficient. It also allows
them more space so they can move around without wanting to die. This
might also help alleviate the lines on the merchant road by moving
people south and west of the building which in turn might increase foot
traffic to the next street over in town.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->