Matthios Dungeon Addition (#839)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## Matthios dungeon map uploads. Two new maps, one is acidmagetower
which was made 30x30 and the other is thelastbreath which is a 45x45
tile dungeon. Both shouldn't need editing to make them work in the rng
pool. The editing to the code is for the rooms to spawn in and it should
be correct with its placement.

![image](https://github.com/user-attachments/assets/535741cb-e2e9-48bd-8f50-8113ee6c09fb)

![image](https://github.com/user-attachments/assets/05e0bcfb-255f-4780-b59e-57ca137f4e58)

![image](https://github.com/user-attachments/assets/db9ffafe-53bf-464c-9cd4-78f6efe1153f)

![image](https://github.com/user-attachments/assets/ed38ede2-0ff7-4674-9fe9-b3ae9180cc0d)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## This adds more maps for the Matthios dungeon that not only increase
the danger but increase the randomization of the dungeon rooms.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>